feat: 바로팟 호스트 컴포넌트 분리
