+++ src/app/features/baropot/components/detail/BaropotDetailAction.tsx
@@ -1,8 +1,8 @@
-// BaropotDetailAction.tsx
-import { useParams, useRouter } from 'next/navigation';
+import { useParams } from 'next/navigation';
 import { BaropotDetailResponse } from '@/app/shared/types/baropots';
 import { motion } from 'framer-motion';
 import { useJoinBaropot } from '@/app/shared/hooks/queries/useBaropot';
+import { useToast } from '@/app/shared/hooks/useToast';
 
 interface BaropotDetailActionProps {
   baropot: BaropotDetailResponse;
@@ -11,23 +11,40 @@ interface BaropotDetailActionProps {
 export default function BaropotDetailAction({
   baropot,
 }: BaropotDetailActionProps) {
-  const router = useRouter();
   const params = useParams();
   const baropotId = Number(params.baropotId);
-
+  const toast = useToast();
   const joinBaropotMutation = useJoinBaropot();
 
   const handleJoinBaropot = () => {
-    // TODO: 검증 케이스 추가 (상태에 맞춰서 입장할 수 있도록 처리)
-    // - OPEN, CLOSED, FULL, CANCELLED, COMPLETED 상태에 맞춰서 처리
-    // TODO: 상태 메세지 추가 모달창 구현
-    if (baropot.status === 'OPEN') {
-      joinBaropotMutation.mutate({
-        baropotId,
-        message: {
-          joinMessage: '바로팟에 참가했습니다!',
-        },
-      });
+    switch (baropot.status) {
+      case 'OPEN':
+        joinBaropotMutation.mutate(
+          {
+            baropotId,
+            message: {
+              joinMessage: '바로팟에 참가했습니다!',
+            },
+          },
+          {
+            onSuccess: () => {
+              toast.success('바로팟에 참가했습니다!');
+            },
+            onError: () => {
+              toast.error('바로팟 참가에 실패했습니다.');
+            },
+          }
+        );
+        break;
+      case 'CANCELLED':
+        toast.error('바로팟이 취소되었습니다.');
+        break;
+      case 'COMPLETED':
+        toast.error('바로팟이 완료되었습니다.');
+        break;
+      case 'FULL':
+        toast.error('바로팟이 가득 찼습니다.');
+        break;
     }
   };
 
@@ -39,17 +56,6 @@ export default function BaropotDetailAction({
       transition={{ delay: 0.3, type: 'spring', stiffness: 100 }}
     >
       <div className="mx-auto flex max-w-lg space-x-4">
-        <motion.button
-          onClick={() => router.push(`/baropot/${baropotId}/edit`)}
-          className="group relative flex-1 overflow-hidden rounded-xl border-2 border-blue-600 px-6 py-4 font-bold text-blue-600 transition-all duration-200 hover:bg-blue-50"
-          whileHover={{ scale: 1.02 }}
-          whileTap={{ scale: 0.98 }}
-        >
-          <span className="relative z-10 flex items-center justify-center">
-            <span>수정하기</span>
-          </span>
-        </motion.button>
-
         <motion.button
           onClick={handleJoinBaropot}
           disabled={baropot.status !== 'OPEN'}
