+++ src/app/features/baropot/components/host/hostClient/HostFilterSection.tsx
@@ -1,0 +1,56 @@
+import { BaropotStatus, baropotStatusKorean } from '@/app/shared/types/enums';
+import { Input } from '@/app/shared/ui';
+import { BaropotsQueries } from '@/app/shared/types/baropots';
+
+export default function HostFilterSection({
+  queries,
+  handleFilterChange,
+}: {
+  queries: BaropotsQueries;
+  handleFilterChange: (key: keyof BaropotsQueries, value: any) => void;
+}) {
+  return (
+    <>
+      <h2 className="mb-4 text-lg font-semibold text-[#2B2B2B]">필터</h2>
+
+      <div className="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3">
+        {/* 상태 필터 */}
+        <div>
+          <label className="mb-2 block text-sm font-medium text-[#2B2B2B]">
+            상태
+          </label>
+          <select
+            value={queries.statusList || ''}
+            onChange={(e) => handleFilterChange('statusList', e.target.value)}
+            className="w-full rounded-lg border border-gray-300 px-3 py-2 focus:border-transparent focus:ring-2 focus:ring-[#1C4E80]"
+          >
+            <option value="">전체</option>
+            {Object.values(BaropotStatus).map((status) => (
+              <option key={status} value={status}>
+                {baropotStatusKorean[status as BaropotStatus]}
+              </option>
+            ))}
+          </select>
+        </div>
+
+        {/* 제목 검색 */}
+        <Input
+          type="text"
+          placeholder="바로팟 제목 검색"
+          value={queries.title || ''}
+          onChange={(e) => handleFilterChange('title', e.target.value)}
+          label="제목 검색"
+        />
+
+        {/* 맛집 이름 검색 */}
+        <Input
+          type="text"
+          placeholder="맛집 이름 검색"
+          value={queries.restaurantName || ''}
+          onChange={(e) => handleFilterChange('restaurantName', e.target.value)}
+          label="맛집 이름"
+        />
+      </div>
+    </>
+  );
+}
